[test] gathering signing tests related to keys
